refactor(settings): extract policy loading into dedicated policy_loader module
Move policy file validation, Starlark evaluation, and compilation logic
from the oversized settings.rs into a new policy_loader module. Settings
now delegates to policy_loader for these operations while retaining its
role as the orchestrator for configuration loading.
